refactor: moving mapping logic for partial authorization from token to token mapper
